+++ sihl/src/web_http.ml
@@ -1,89 +1,30 @@
-open Contract_http
-
-let log_src = Logs.Src.create ("sihl.service." ^ Contract_http.name)
-
-module Logs = (val Logs.src_log log_src : Logs.LOG)
-
-let get path handler = Get, path, handler
-let post path handler = Post, path, handler
-let put path handler = Put, path, handler
-let delete path handler = Delete, path, handler
-let any path handler = Any, path, handler
-
-let router ?(scope = "/") ?(middlewares = []) routes =
-  { scope; routes; middlewares }
-;;
-
-let trailing_char s =
-  let length = String.length s in
-  try Some (String.sub s (length - 1) 1) with
-  | _ -> None
-;;
-
-let tail s =
-  try String.sub s 1 (String.length s - 1) with
-  | _ -> ""
-;;
-
-let prefix prefix (meth, path, handler) =
-  let path =
-    match trailing_char prefix, Astring.String.head path with
-    | Some "/", Some '/' -> Printf.sprintf "%s%s" prefix (tail path)
-    | _, _ -> Printf.sprintf "%s%s" prefix path
-  in
-  meth, path, handler
-;;
-
-let apply_middleware_stack middleware_stack (meth, path, handler) =
-  (* The request goes through the middleware stack from top to bottom, so we
-     have to reverse the middleware stack *)
-  let middleware_stack = List.rev middleware_stack in
-  let wrapped_handler =
-    List.fold_left
-      (fun handler middleware -> Rock.Middleware.apply middleware handler)
-      handler
-      middleware_stack
-  in
-  meth, path, wrapped_handler
-;;
-
-let router_to_routes { scope; routes; middlewares } =
-  routes
-  |> List.map (prefix scope)
-  |> List.map (apply_middleware_stack middlewares)
-;;
-
-let externalize_path ?prefix path =
-  let prefix =
-    match prefix, Core_configuration.read_string "PREFIX_PATH" with
-    | Some prefix, _ -> prefix
-    | _, Some prefix -> prefix
-    | _ -> ""
-  in
-  path
-  |> String.split_on_char '/'
-  |> List.cons prefix
-  |> String.concat "/"
-  |> Stringext.replace_all ~pattern:"//" ~with_:"/"
-;;
+include Contract_http
 
 let to_opium_builder (meth, path, handler) =
-  let open Contract_http in
+  let open Web in
   match meth with
   | Get -> Opium.App.get path handler
+  | Head -> Opium.App.head path handler
+  | Options -> Opium.App.options path handler
   | Post -> Opium.App.post path handler
+  | Patch -> Opium.App.patch path handler
   | Put -> Opium.App.put path handler
   | Delete -> Opium.App.delete path handler
   | Any -> Opium.App.all path handler
 ;;
 
 let routers_to_opium_builders routers =
+  let open Web in
   routers
   |> List.map (fun router ->
-         let routes = router_to_routes router in
+         let routes = routes_of_router router in
          routes |> List.map to_opium_builder |> List.rev)
   |> List.concat
 ;;
+
+let log_src = Logs.Src.create ("sihl.service." ^ Contract_http.name)
+
+module Logs = (val Logs.src_log log_src : Logs.LOG)
 
 type config = { port : int option }
 
@@ -94,7 +35,7 @@ let schema =
   make [ optional (int ~default:3000 "PORT") ] config
 ;;
 
-let registered_routers = ref []
+let registered_router = ref None
 let registered_middlewares = ref []
 let started_server = ref None
 
@@ -110,7 +51,12 @@ let start_server () =
      every single request goes through them, not only requests that match
      handlers *)
   let app = List.fold_left (fun app builder -> builder app) app middlewares in
-  let builders = routers_to_opium_builders !registered_routers in
+  let router =
+    match !registered_router with
+    | None -> raise @@ Exception "No router registered"
+    | Some router -> router
+  in
+  let builders = routers_to_opium_builders [ router ] in
   let app = List.fold_left (fun app builder -> builder app) app builders in
   (* We don't want to block here, the returned Lwt.t will never resolve *)
   let* server = Opium.App.start app in
@@ -143,8 +89,8 @@ let stop () =
 
 let lifecycle = Core_container.create_lifecycle "http" ~start ~stop
 
-let register ?(middlewares = []) ?(routers = []) () =
-  registered_routers := routers;
+let register ?(middlewares = []) router =
+  registered_router := Some router;
   registered_middlewares := middlewares;
   let configuration = Core_configuration.make ~schema () in
   Core_container.Service.create